requirements.txt
python-nmap>=0.7.1
shodan>=1.28.0
vulners>=2.0.2
securitytrails>=0.1.2
rich>=13.3.0
requests>=2.28.0
jinja2>=3.1.0
